requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
apispec==6.6.1
attrs==24.2.0
bech32==1.2.0
bidict==0.23.1
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cosmpy==0.9.2
dataclasses-json==0.6.7
deepgram-sdk==3.7.3
deprecation==2.1.0
distro==1.9.0
docutils==0.21.2
ecdsa==0.19.0
fastapi==0.115.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
grpcio==1.66.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.4.1
lazy_loader==0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.4
networkx==3.4.1
nh3==0.2.18
numpy==2.1.2
openai==1.52.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-multipart==0.0.12
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
redis==5.1.1
referencing==0.35.1
reflex==0.6.3
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.13
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.14.1
setuptools==75.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.40.0
starlette-admin==0.14.1
streamlit==1.39.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tokenizers==0.20.1
toml==0.10.2
tomlkit==0.13.2
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
twine==5.1.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uagents==0.16.2
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.14.0
zipp==3.20.2
